Input checks for cashier.cpp
Added check for char in addBookMenu().

Added option to exit from entering an identifier to cart. Previously you
would get stuck in the menu with no way to get back to the main menu
other than closing the program. This is currently set to 0 but it can be
changed to any value you guys want. It's all I could think of since the
program already tracks -1 for no ID found. I don't believe 0 is ever
used as an identifier since the default constructor sets it to 1. Change
it if I'm wrong please!

Changed some output text