Check if database is full
Change location of checking for a full database to prevent recursion.
This is different from how it was before. The previous version would
output the message after all the books up to size 1024 were created.
This version checks before any books are created to see if database is
full. It still functions the same for the user.